Methods Rework Partial Implementation
