+++ src/board/control/gameStartListeners.js
@@ -1,0 +1,8 @@
+import gameStart from './functions/gameStart';
+
+export default (target) => {
+  target.disabled = true;
+  target.addEventListener('click', () => {
+    gameStart();
+  });
+};
